added support for ownership based access control and some refactor
