+++ server/middlewares/injectRequestId/index.js
@@ -5,11 +5,9 @@ function RequestIdInjectedLogger(options, requestId) {
     this.log = options.log.child({ requestId });
 }
 
-const injectRequestId = () => (_request, _response, next) => {
+export const injectRequestId = () => (_request, _response, next) => {
     const requestId = uuid();
     const logger = new RequestIdInjectedLogger({ log }, requestId);
     global.log = logger.log;
     next();
 };
-
-module.exports = injectRequestId;
