+++ server/utils/logger.js
@@ -1,6 +1,6 @@
 import bunyan from 'bunyan';
 
-module.exports = global.log = bunyan.createLogger({
+export default global.log = bunyan.createLogger({
     name: 'parcel-node-mongo-express',
     streams: [
         {
